Change physical dimensions for 800x480 screen
In the code there are hardcoded dimensions based on resolution:
for 800x480 screen the reported resolution is 800x480 and dimensions
140x80 (which have different aspect ratio: 7:4 instead of 5:3), that
leads to actual resolution becoming 840x480 (480*140/80=840);
for 1024x600 screen the reported resolution is 960x560 and dimensions
225x125 mm (9:6 instead of 128:75), that should lead to actual
resolution becoming 1008x560;
for all other resolutions the physical dimensions are set as width/6 and
height/6.

This patch sets dimensions for 800x480 screen to 110x66 mm, which is
very close to actual physical size of the screen and gives exactly
native resolution and aspect ratio.

I didn't change the physical dimensions and resolution for 1024x600
screen, because I don't have such device (and this firmware is probably
not suitable for it).